Move call site linking in extra pass.
The call site linking needs already linked type information and thus
needs to go in a separated pass which runs after the Linker.
