+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/passes/linking/linker/Linker.scala
@@ -33,8 +33,6 @@ class Linker(cpg: Cpg) extends CpgPass(cpg) {
 
     linkAstChildToParent(dstGraph)
 
-    linkCalls(dstGraph)
-
     linkToSingle(
       srcLabels = List(NodeTypes.TYPE),
       dstNodeLabel = NodeTypes.TYPE_DECL,
@@ -257,28 +255,6 @@ class Linker(cpg: Cpg) extends CpgPass(cpg) {
       "Could not create edge. Source lookup failed. " +
         s"edgeType=$edgeType, srcNodeType=$srcNodeType, srcFullName=$srcFullName, " +
         s"dstNodeType=$dstNodeType, dstNodeId=$dstNodeId")
-  }
-
-  private def linkCalls(dstGraph: DiffGraph): Unit = {
-    cpg.call.sideEffect { call =>
-      linkCall(call, dstGraph)
-    }.exec()
-  }
-
-  private def linkCall(call: nodes.Call, dstGraph: DiffGraph): Unit = {
-    val receiver = call.vertices(Direction.OUT, EdgeTypes.RECEIVER).next
-    val receiverTypeDecl = receiver.vertices(Direction.OUT, EdgeTypes.EVAL_TYPE).next
-      .vertices(Direction.OUT, EdgeTypes.REF).next
-
-    val resolvedMethodOption =
-      receiverTypeDecl.vertices(Direction.OUT, EdgeTypes.BINDS).asScala.collectFirst {
-        case binding: nodes.Binding if binding.name == call.name && binding.signature == call.signature =>
-          binding.vertices(Direction.OUT, EdgeTypes.REF).nextChecked.asInstanceOf[nodes.Method]
-      }
-
-    resolvedMethodOption.foreach { method =>
-      dstGraph.addEdgeInOriginal(call, method, EdgeTypes.CALL)
-    }
   }
